django_project/requirements.txt
asgiref==3.6.0
autopep8==2.0.2
backports.zoneinfo==0.2.1
bpyutils==0.5.8
CacheControl==0.12.11
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==39.0.2
Django==2.1.15
django-ckeditor==6.5.1
django-crispy-forms==2.0
django-js-asset==2.0.0
django-model-utils==4.3.1
django-mssql-backend==2.8.1
django-pyodbc==1.1.3
django-pyodbc-azure==2.1.0.0
djangorestframework==3.14.0
firebase-admin==6.1.0
google-api-core==2.11.0
google-api-python-client==2.82.0
google-auth==2.16.2
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.10.0
google-cloud-storage==2.7.0
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.59.0
greenlet==2.0.2
grpcio==1.51.3
grpcio-status==1.51.3
httplib2==0.22.0
idna==3.4
msgpack==1.0.5
mssql-django==1.2
numpy==1.24.2
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
pip-review==1.3.0
pipupgrade==1.12.0
proto-plus==1.22.2
protobuf==4.22.1
psycopg2==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pycparser==2.21
PyJWT==2.6.0
pyodbc==4.0.35
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7.1
requests==2.28.2
rsa==4.9
six==1.16.0
SQLAlchemy==2.0.7
sqlparse==0.4.3
toml==0.10.2
tomli==2.0.1
typing_extensions==4.5.0
tzdata==2022.7
Unipath==1.1
uritemplate==4.1.1
urllib3==1.26.15
